requirements.txt
aiofiles==25.1.0
aiogram==3.27.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
attrs==26.1.0
audioop-lts==0.2.2
audioread==3.1.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
decorator==5.2.1
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.47.0
magic-filter==1.0.12
MarkupSafe==3.0.3
msgpack==1.1.2
multidict==6.7.1
numba==0.65.0
numpy==2.4.4
packaging==26.1
platformdirs==4.9.6
pooch==1.9.0
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.49
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
yarl==1.23.0
